% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{boot_mean_ci}
\alias{boot_mean_ci}
\title{Calculate mean and bootstrapped confidence limits using the bias corrected and
accelarated (BCa) method, which is more robust with low sample numbers and
non-normal distributions of bootstrapped means.}
\usage{
boot_mean_ci(df, value, reps, conf, ...)
}
\arguments{
\item{df}{Dataframe containing grouping and value fields.}

\item{value}{Value field for statistics.}

\item{reps}{Number of bootstraping repetitions. May need to increase if
BCa error occurs.}

\item{conf}{Confidence limits probability (e.g., 0.95).}

\item{...}{One or more grouping columns.}
}
\value{
Dataframe with mean and confidence limits.
}
\description{
Calculate mean and bootstrapped confidence limits using the bias corrected and
accelarated (BCa) method, which is more robust with low sample numbers and
non-normal distributions of bootstrapped means.
}
\examples{
df <- data.frame(
x = rlnorm(n = 100, meanlog = 1.58, sdlog = 1.09),
d = sample(c("a", "b"), 100, replace = TRUE)
)

boot_mean_ci(df, x, reps = 100, conf = 0.9, d)

}
\author{
Alex M Trueman, 2018-05-08

Based on https://www.painblogr.org/2017-10-18-purrring-through-bootstraps
by Peter Kamerman (@painblogR)
}
