% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{interval_mid}
\alias{interval_mid}
\title{Extract the midpoint of an interval assigned to a data frame.}
\usage{
interval_mid(x, dp = 1)
}
\arguments{
\item{x}{is a data-frame column with intervals typically applied using the
base-r `cut` function.}

\item{dp}{is the number of decimal places for the base-r `round` function.}
}
\value{
an atomic vector of type double.
}
\description{
The mid-point is used as an attractive label for plot axes.
}
\author{
Alex Trueman
}
