% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{swath_data}
\alias{swath_data}
\title{Generate means by geographical slices (swaths).}
\usage{
swath_data(df, value, group, axis, slices)
}
\arguments{
\item{df}{Data frame with at least one coordinate field.}

\item{value}{Grade column being evaluated.}

\item{group}{Grouping column (e.g., domain).}

\item{axis}{Axis column for slicing (e.g., x).}

\item{slices}{List of slice interval coordinates as numeric vector.
Can be generated from function \code{swath_seq}.}
}
\value{
data frame
}
\description{
For a given set of coordinate intervals calculate the mean of the data
within each interval (swath).
}
\author{
Alex M Trueman
}
