% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{swath_data_ci}
\alias{swath_data_ci}
\title{Generate means by geographical slices (swaths).
Extended to include lower and upper confidence limits by bootstrapping.}
\usage{
swath_data_ci(df, value, group, axis, slices, reps = 10000, conf = 0.95)
}
\arguments{
\item{df}{Data frame with at least one coordinate field.}

\item{value}{Column name to be evaluated (e.g., au).}

\item{group}{Grouping column (e.g., domain).}

\item{axis}{Axis column for slicing (e.g., x).}

\item{slices}{List of slice starting coordinates as numeric vector.
Can be generated from function \code{swath_seq}.}

\item{reps}{(default = 10000) Number of bootstrap repetitions.}

\item{conf}{(default = 0.95) Confidence level.}
}
\value{
dataframe
}
\description{
For a given set of coordinate intervals calculate the mean of the data
within each interval (swath).
}
\author{
Alex M Trueman
}
