% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{swath_seq}
\alias{swath_seq}
\title{Create a sequence of swath intervals for a given axis and increment size.}
\usage{
swath_seq(axis, slice_size, block_inc)
}
\arguments{
\item{axis}{is a vector of single axis coordinates (e.g., model$x).}

\item{slice_size}{is the swath slice size for the axis.}

\item{block_inc}{is the size of parent blocks in the model used for rounding.}
}
\value{
a numeric vecotor with a sequence of slice coordinates.
}
\description{
Create a sequence of swath intervals for a given axis and increment size.
}
\author{
Alex Trueman
}
