% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swath.r
\name{swath_seq}
\alias{swath_seq}
\title{Create a sequence of swath intervals for a given axis and increment size.}
\usage{
swath_seq(axis, slice_size, block_inc)
}
\arguments{
\item{axis}{numeric vector of single axis coordinates (e.g., model$x).}

\item{slice_size}{numeric swath slice size for the axis.}

\item{block_inc}{numeric size of parent blocks in the model.
The minimum and maximum coordinate of slicing is rounded to the
nearest \code{block_inc}.}
}
\value{
numeric vector with a sequence of slice coordinates.
}
\description{
Swath statistics are generated for slices through the spatial data.
\code{swath_seq} generates slice intervals based on some input spatial
data, a block size, and slice width.
'
}
\author{
Alex Trueman
}
